Commonize `ktor-server-csrf`
